use latest hibernate version

+++ struts2-jquery-grid-showcase/src/main/java/com/jgeppert/struts2/jquery/grid/showcase/dao/CustomersDao.java
@@ -44,7 +44,7 @@ public class CustomersDao extends AbstractSimpleGenericDao<Customers, Integer> {
     {
       Criteria criteria = dc.getExecutableCriteria(hSession);
       criteria.setProjection(Projections.rowCount());
-      return ((Integer) criteria.list().get(0)).intValue();
+      return ((Long) criteria.list().get(0)).intValue();
     }
     catch (HibernateException e)
     {
